Shift PreviewsState usage in ResourceResponseInfo to an int.
This breaks the dependency of ResourceResponseInfo on PreviewsState and allows
a (future) move into //services/network.

Bug: 753658

+++ content/browser/loader/resource_loader.cc
@@ -72,7 +72,8 @@ void PopulateResourceResponse(
   const content::ResourceRequestInfo* request_info =
       content::ResourceRequestInfo::ForRequest(request);
   if (request_info)
-    response->head.previews_state = request_info->GetPreviewsState();
+    response->head.previews_state =
+        static_cast<int>(request_info->GetPreviewsState());
   if (info->ShouldReportRawHeaders()) {
     response->head.raw_request_response_info = BuildRawRequestResponseInfo(
         *request, raw_request_headers, raw_response_headers);
